refactored anaire methods in separated files
