+++ src/cloud_anaire.hpp
@@ -1,0 +1,6 @@
+#define ANAIRE_HOST "mqtt.anaire.org"
+#define ANAIRE_TOPIC "measurement"
+#define ANAIRE_PORT 80
+
+void anaireLoop ();
+void anaireInit ();
